refactor: move Interests component hooks to hooks.ts file
