+++ src/components/interests/hooks.ts
@@ -1,0 +1,14 @@
+import { create } from "zustand";
+
+export type State = {
+  interests: string;
+};
+
+type Action = {
+  updateInterests: (description: State["interests"]) => void;
+};
+
+export const useStore = create<State & Action>()((set) => ({
+  interests: "",
+  updateInterests: (interests) => set(() => ({ interests: interests })),
+}));
